# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.PaginatedRolesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/role.RoleResponse'
        type: array
      pagination:
        $ref: '#/definitions/internal_feature_role_handler.Pagination'
    type: object
...
  internal_feature_role_handler.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  internal_feature_role_handler.Response:
    properties:
      data: {}
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  permission.AssignPermissionRequest:
    properties:
      permission_id:
        type: string
    type: object
...
  permission.PermissionResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  role.AssignRoleRequest:
    properties:
      role_id:
        type: string
    type: object
  role.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  role.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  role.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A Golang starter project with Vertical Slice Architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Golang Starter API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get a paginated list of roles
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_role_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PaginatedRolesResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_role_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get a role by their ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_role_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update a role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_role_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/role.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - roles
  /roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: Get permissions for a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_role_handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/permission.PermissionResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Get role permissions
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Assign a permission to a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/permission.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Assign permission to role
      tags:
      - roles
  /roles/{id}/permissions/{permissionId}:
    delete:
      consumes:
      - application/json
      description: Revoke a permission from a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Revoke permission from role
      tags:
      - roles
...
  /users/{id}/roles:
    get:
      consumes:
      - application/json
      description: Get roles for a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_feature_role_handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/role.RoleResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Get user roles
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Assign a role to a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/role.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Assign role to user
      tags:
      - users
  /users/{id}/roles/{roleId}:
    delete:
      consumes:
      - application/json
      description: Revoke a role from a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Role ID
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_feature_role_handler.Response'
      security:
      - BearerAuth: []
      summary: Revoke role from user
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"